Add --keep_images flag

+++ files.py
@@ -86,7 +86,7 @@ def write_location_mapping_script(locations: list[dict[str, any]], location_name
         file.write(location_to_id_string + id_to_location_string)
 
 
-def copy_default_files(items: list[dict[str, any]], options: list[str], pack_root: str):
+def copy_default_files(items: list[dict[str, any]], options: list[str], keep_images: bool, pack_root: str):
     if not os.path.isabs(pack_root):
         raise SyntaxError(f"Pack root must be an absolute path! Given pack root: {pack_root}")
     item_images_dirpath: str = os.path.join(pack_root, "images/items/")
@@ -98,14 +98,16 @@ def copy_default_files(items: list[dict[str, any]], options: list[str], pack_roo
             continue
         image_filename = f"{to_snake_case(item['name'])}.png"
         item_image_filepath = os.path.join(item_images_dirpath, image_filename)
-        shutil.copyfile("./data/default_item_option_image.png", item_image_filepath)
+        if not os.path.exists(item_image_filepath) or not keep_images:
+            shutil.copyfile("./data/default_item_option_image.png", item_image_filepath)
     option_images_dirpath: str = os.path.join(pack_root, "images/options/")
     if not os.path.exists(option_images_dirpath):
         os.makedirs(option_images_dirpath)
     for option in options:
         image_filename = f"{option}.png"
         item_image_filepath = os.path.join(option_images_dirpath, image_filename)
-        shutil.copyfile("./data/default_item_option_image.png", item_image_filepath)
+        if not os.path.exists(item_image_filepath) or not keep_images:
+            shutil.copyfile("./data/default_item_option_image.png", item_image_filepath)
     ap_lua_filepath = os.path.join(pack_root, "scripts/archipelago/archipelago.lua")
     if not os.path.exists(os.path.dirname(ap_lua_filepath)):
         os.makedirs(os.path.dirname(ap_lua_filepath))
@@ -121,7 +123,8 @@ def copy_default_files(items: list[dict[str, any]], options: list[str], pack_roo
     placeholder_map_image_filepath = os.path.join(pack_root, "images/maps/main.png")
     if not os.path.exists(os.path.dirname(placeholder_map_image_filepath)):
         os.makedirs(os.path.dirname(placeholder_map_image_filepath))
-    shutil.copyfile("./data/placeholder_map.png", placeholder_map_image_filepath)
+    if not os.path.exists(placeholder_map_image_filepath) or not keep_images:
+        shutil.copyfile("./data/placeholder_map.png", placeholder_map_image_filepath)
 
 
 def write_common_lua_scripts(item_groups: dict[str, list[str]], pack_root: str):
